Comment Vote System
Also sorts comments based on most popular

+++ src/server/api/vote.js
@@ -3,24 +3,76 @@ const router = express.Router();
 const {PrismaClient} = require("@prisma/client");
 const prisma = new PrismaClient();
 
-router.delete("/:id",  async (req, res, next) =>{
+router.post('/', require('../auth/middleware'), async (req, res, next) => {
 
-    const vote = await prisma.vote.getFirst({
-        where:{
-            userId: req.body.userId,
-            submissionId: req.params.id
+    try{
+        const exist = await prisma.vote.findFirst({
+            where:{
+                userId:  Number(req.body.userId),
+                commentId:  Number(req.body.commentId),
+            }
+        })
+
+        if(exist){
+
+            await prisma.vote.delete({
+                where: {
+                    id: exist.id,
+                    userId:  Number(req.body.userId),
+                    commentId:  Number(req.body.commentId),
+                }
+            });
+            if(exist.type!==req.body.type){
+                await prisma.vote.create({
+                    data: {
+                        userId:  Number(req.body.userId),
+                        commentId:  Number(req.body.commentId),
+                        type: req.body.type
+                    }
+                });
+            }
+        }else{
+            await prisma.vote.create({
+                data: {
+                    userId:  Number(req.body.userId),
+                    commentId:  Number(req.body.commentId),
+                    type: req.body.type
+                }
+            });
         }
-    })
 
-    const deleteVote = await prisma.vote.delete({
-        where:{
-            id:vote.id
-        }
-    })
+        const comment =  await prisma.comment.findFirst({
+            where:{
+                id:req.body.commentId
+            }
+        })
 
-    res.send(deleteVote);
+        const finalPost = await prisma.post.findFirst({
+            where: {
+                id: comment.postId
+            },
+            include: {
+                post_tag: {
+                    include: {
+                        tag: true
+                    }
+                },
+                author: true,
+                like: true,
+                comment: {
+                    include:{
+                        author:true,
+                        vote:true
+                    }
+                }
+            }
+        })
+        res.send(finalPost)
 
+
+    }catch(err){
+        next(err)
+    }
 })
 
-
-exports.module = router;
+module.exports = router;
